Projects and fixes for VC++ 6, 2022

+++ 2024/c/day08.c
@@ -22,7 +22,7 @@ main(int argc, char **argv)
 	assert(feof(stdin));
 	lf = strchr(g[0], '\n');
 	assert(lf);
-	w = lf - g[0];
+	w = (int)(lf - g[0]);
 
 	/*
 	 * Find antenna pairs, then project backwards from the first,
